improve performance
The module was tested by finding scrobbles of "Conspiracy" by King
Diamond in my exported Last.fm data (246249 scrobbles). Processing took
about 25.5 seconds.

It turns out that most of the time was spent in asciify(). First of all,
three calls to translate() are very expensive. Using one call with
bigger translation map is more efficient: processing time is decreased
to about 16 seconds.

Another improvement is avoiding the translation altogether if it's not
needed. Checking if string contains typographic characters and returning
the same string immediately if it's not the case decreases processing
time to about 9 seconds.

Changing nesting of loops in is_typographic() decreases processing time
to about 8.3 seconds. Probably because the character distribution is
such that the first typographic character is now found earlier, and the
loop exits faster.

Finally, checking for typographic strings in compare_normalized() avoids
an unnecessary call to normalize(), which reduces processing time to
about 7.5 seconds.

Overall performance gain is about 3.4x.