rename cmp_fuzzy_normalized_numbers()
The new name cmp_normalized_fuzzy() fits the other comparison function.
There is no need to indicate numbers normalization in the name, because
it's necessary in fuzzy matching (i.e. there will be no function for
fuzzy matching which doesn't internally replace numbers).

This is a breaking change which needs adaptation in clients.