provide a wrapper around fuzzy matching function
cmp_fuzzy() is a thin wrapper around jaro_winkler_similarity() from the
jellyfish package. Implement cmp_normalized_fuzzy() using cmp_fuzzy().

The idea is to centralize the use of a third-party function in a single
location. This will make it easier to change the fuzzy matching method
consistently across the code.

This could be useful for #1.